@@ -83,15 +83,16 @@
         "arrowParens": "avoid"
     },
     "dependencies": {
-        "@angular/animations": "~8.2.14",
-        "@angular/common": "~8.2.14",
-        "@angular/compiler": "~8.2.14",
-        "@angular/core": "~8.2.14",
-        "@angular/forms": "~8.2.14",
-        "@angular/platform-browser": "~8.2.14",
-        "@angular/platform-browser-dynamic": "~8.2.14",
-        "@angular/platform-server": "~8.2.14",
-        "@angular/router": "~8.2.14",
+        "@angular/animations": "~9.1.12",
+        "@angular/common": "~9.1.12",
+        "@angular/compiler": "~9.1.12",
+        "@angular/core": "~9.1.12",
+        "@angular/forms": "~9.1.12",
+        "@angular/localize": "9",
+        "@angular/platform-browser": "~9.1.12",
+        "@angular/platform-browser-dynamic": "~9.1.12",
+        "@angular/platform-server": "~9.1.12",
+        "@angular/router": "~9.1.12",
         "@ctrl/ngx-codemirror": "^3.0.0",
         "@fortawesome/angular-fontawesome": "^0.5.0",
         "@fortawesome/fontawesome-svg-core": "^1.2.26",
@@ -114,22 +115,22 @@
         "snapsvg": "^0.5.1",
         "stream": "^0.0.2",
         "tslib": "^1.10.0",
-        "zone.js": "~0.9.1"
+        "zone.js": "~0.10.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.803.28",
-        "@angular/cli": "~8.3.28",
-        "@angular/compiler-cli": "~8.2.14",
-        "@angular/language-service": "~8.2.14",
+        "@angular-devkit/build-angular": "~0.901.11",
+        "@angular/cli": "~9.1.11",
+        "@angular/compiler-cli": "~9.1.12",
+        "@angular/language-service": "~9.1.12",
         "@commitlint/cli": "^8.3.4",
         "@commitlint/config-angular": "^8.3.4",
         "@compodoc/compodoc": "^1.1.11",
         "@types/d3": "^5.7.2",
         "@types/jasmine": "~3.3.8",
         "@types/jasminewd2": "~2.0.3",
-        "@types/node": "~10.17.16",
+        "@types/node": "^12.11.1",
         "angular-cli-ghpages": "^0.6.2",
-        "codelyzer": "^5.2.1",
+        "codelyzer": "^5.1.2",
         "conventional-recommended-bump": "^6.0.5",
         "cross-var": "^1.1.0",
         "gzipper": "^3.4.2",
